+++ src/RenderBoard.ts
@@ -1,6 +1,6 @@
-import { Board } from './Board'
+import { Board } from "./Types";
 
-export const RED_MARBLE: String= `\u001b[31m \u2022 \u001b[0m`;
+export const RED_MARBLE: String = `\u001b[31m \u2022 \u001b[0m`;
 export const BLUE_MARBLE: String = `\u001b[34m \u2022 \u001b[0m`;
 export const WHITE_MARBLE: String = `\u001b[37m \u2022 \u001b[0m`;
 export const EMPTY_MARBLE: String = `   `;
@@ -9,28 +9,27 @@ const MARBLE_COLORS = [EMPTY_MARBLE, RED_MARBLE, BLUE_MARBLE, WHITE_MARBLE];
 const alphabet: String = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
 
 export function renderBoard(board: Board): String {
-    const columnLetters = alphabet.substr(0, board[0].length)
-    const header = columnLetters.split("").map(char => ` ${char} `)
-    const firstLine = '   ' + header.join("")
+  const columnLetters = alphabet.substr(0, board[0].length);
+  const header = columnLetters.split("").map((char) => ` ${char} `);
+  const firstLine = "   " + header.join("");
 
-    let result = firstLine + '\n';
+  let result = firstLine + "\n";
 
-    for (let i = 0; i < board.length; i++) {
-        let line = ` ${i} `;
-        for (const marble of board[i]) {
-            line += marbleValuetoANSIColorCode(marble);
-
-        }
-        result += line + '\n';
+  for (let i = 0; i < board.length; i++) {
+    let line = ` ${i} `;
+    for (const marble of board[i]) {
+      line += marbleValuetoANSIColorCode(marble);
     }
+    result += line + "\n";
+  }
 
-    return result;
+  return result;
 }
 
 function marbleValuetoANSIColorCode(marble: number): String {
-    return MARBLE_COLORS[marble];
+  return MARBLE_COLORS[marble];
 }
 
 export function renderToConsole(graphicalBoard: String) {
-    console.log(graphicalBoard)
+  console.log(graphicalBoard);
 }
