+++ src/index.ts
@@ -1,9 +1,9 @@
 import { getInitialBoard, canMoveMarbleInDirection } from "./Board";
 import { renderBoard, renderToConsole } from "./RenderBoard";
 import { boardToGraph } from "./Graph";
-import { Graph } from "./Types";
+import { Graph, Board } from "./Types";
 
-const firstBoard = getInitialBoard();
+const firstBoard: Board = getInitialBoard();
 const graphicalBoard = renderBoard(firstBoard);
 
 renderToConsole(graphicalBoard);
@@ -12,12 +12,14 @@ let graph: Graph = boardToGraph(firstBoard);
 if (graph) {
   console.log(graph);
 
-  let moveAuth = canMoveMarbleInDirection(graph, "0,0", "S");
-  console.log(moveAuth);
-  let moveAuth2 = canMoveMarbleInDirection(graph, "0,0", "N");
-  console.log(moveAuth2);
-  let moveAuth3 = canMoveMarbleInDirection(graph, "0,0", "W");
-  console.log(moveAuth3);
-  let moveAuth4 = canMoveMarbleInDirection(graph, "0,0", "E");
-  console.log(moveAuth4);
+  let newMove = canMoveMarbleInDirection(graph, "0,0", "S");
+  console.log(
+    `the move from 0,0 in the south direction is authorized ? `,
+    newMove
+  );
+  let newMove2 = canMoveMarbleInDirection(graph, "0,0", "N");
+  console.log(
+    `the move from 0,0 in the south direction is authorized ? `,
+    newMove2
+  );
 }
